feat(neovim): add nvim-navic plugin
This commit adds the nvim-navic plugin to the neovim configuration. The
plugin is added to the list of plugins in the neovim-overlay.nix file.
Additionally, the setup for nvim-navic is added to the lsp.lua file,
with auto_attach set to true and highlight enabled. The winbar option
is also set to display the location provided by nvim-navic.