feat(neovim): add lz.n plugin and refactor plugins list
This commit includes two main changes. First, it adds the lz.n plugin
to the list of Neovim plugins. This plugin is added to the
unmanaged-plugins list.

Second, it refactors the plugins list to include an 'optional' field.
This field is set to false for all plugins, indicating that they are
not optional and must be included in the Neovim configuration.

A new file, init.lua, is also created in the user/plugin directory
to load the lz.n plugin.