feat(barbar.nvim): Auto hide tabline and disable in dump filetype
This commit makes two changes related to the barbar.nvim plugin. First,
it enables the auto_hide feature, which automatically hides the tabline
when there's only one buffer open. This is done by adding the
"auto_hide" option to the barbar setup call.

Second, it disables the barbar plugin entirely when the filetype is
"dump". This is done by removing the BarbarDisable command and the
showtabline option setting from the dump filetype plugin script.