feat(neovim): Improve filetype detection and plugin settings
This commit improves the filetype detection by changing the 'dump'
filetype to be detected by pattern rather than extension. It also
modifies the settings for the 'dump' filetype to disable the number,
relative number, sign column, winbar, and last status options.

The barbar_nvim plugin settings are updated to disable auto hiding
and exclude the 'dump' filetype.

The lsp plugin settings are updated to create an autocmd for LspAttach
event. This autocmd attaches the nvim-navic to the lsp client and
updates the winbar with the location information from nvim-navic.
`