feat(neovim): update lz.n plugin and openingh-nvim settings
This commit updates the commit hash of the lz.n plugin in the
neovim-overlay.nix file, pointing to a newer version of the plugin.

In addition, the naming convention for the openingh-nvim plugin in the
openingh-nvim.lua file has been updated from "openingh-nvim" to
"openingh.nvim". The event trigger has also been replaced with specific
key bindings for normal and visual modes.