refactor(plugins): reorganize plugins into categories
This commit reorganizes the Neovim plugins into separate files based
on their categories. The categories include 'languages', 'ui', 'git',
and 'utils'. This change makes it easier to manage and understand the
purpose of each plugin. It also improves the readability of the code
and makes it easier to add or remove plugins in the future.